chore:채팅 컴포넌트 분리 및 수정 중에 git pull 최신화
