+++ src/api/supabase.api.ts
@@ -514,3 +514,5 @@ export const getLikesProductsListHandler = async (userId: string) => {
   if (error) throw error;
   return data;
 };
+
+
